refactor: rename util files dateUtils.ts to date.ts and currencyUtils.ts to currency.ts and update their import location
